Updates for updated version

+++ src/edit_reply.php
@@ -3,27 +3,32 @@
 require('includes/application_top.php');
 
 $error = "";
+$replyid = 0;
 
-if ( isset( $_GET['id'] ) || isset($_POST['id']) ) {
+$msgs = array();
+$reply = array();
+
+if ( !isset( $_GET['id'] ) && !isset($_POST['id']) ) {
+	$msgs[] = array('state' => 'nok', 'text' => DELETE_REPLY_NO_ID);
+} else {
 	
-	$replyid = 0;
-	if ( isset($_GET['id']) )
-		$replyid = intval($_GET['id']);
-	else if ( isset($_POST['id']) )
-		$replyid = intval($_POST['id']);
+	$replyid = isset($_GET['id']) ? (int)$_GET['id'] : (int)$_POST['id'];
 		
 	$query = $db->query("select * from elo_reply where reply_id='".$replyid."'");
 	
-	if ( $db->num_rows($query) ) {
+	if ( $db->num_rows($query) < 1 ) {
+		$msgs[] = array('state' => 'nok', 'text' => DELETE_REPLY_NO_RIGHTS);
+	} else {
 		$res = $db->fetch_array($query);
+		$reply = $res;
 
 		if ( ($res['user_id'] == $user_res['user_id'] && $res['reply_date'] > ($time - $conf['max_edit_time']) ) || in_array('IS_ADMIN',$user_rights)) {
 			$topicid = $res['topic_id'];
 	
 			// music sheets
 			$query_m = $db->query("select a.*, ra.rm_id from elo_music as a, elo_reply_music as ra where ra.reply_id='".$replyid."' and ra.music_id=a.music_id");
-					
-/* *************************** */
+			
+			// submit of edited data
 			if ( isset($_POST['edit_topic_admin']) && in_array('IS_ADMIN',$user_rights)) {	
 				$db->query("delete from elo_topic_user where topic_id='".$topicid."'");			
 				if ( isset( $_POST['t_user'] )) {
@@ -56,7 +61,6 @@ if ( isset( $_GET['id'] ) || isset($_POST['id']) ) {
 					$_POST['text'] = htmlentities($_POST['text']);
 							
 				$db->query("update elo_reply set reply_text='".addslashes($_POST['text'])."' where reply_id='".$replyid."'");
-				$reply_id = $replyid;
 				
 				if ( in_array('CREATE_ATTACHMENTS', $user_rights) && isset($_FILES['t_file']) && strlen($_FILES['t_file']['name'])) {
 					processAttachment();
@@ -69,7 +73,7 @@ if ( isset( $_GET['id'] ) || isset($_POST['id']) ) {
 				if ( isset($_POST['noref']) ) {
 					echo "Reply saved.";
 				} else {
-					header("Location: topic.php?id=".$res['topic_id']."#".$reply_id);	
+					header("Location: topic.php?id=".$res['topic_id']."#".$replyid);	
 					exit();
 				}
 			}
@@ -79,220 +83,38 @@ if ( isset( $_GET['id'] ) || isset($_POST['id']) ) {
 			
 			// attachments
 			$query_a = $db->query("select ra.ra_id, a.* from elo_attachment as a, elo_reply_attachment as ra where ra.reply_id='".$replyid."' and ra.attachment_id=a.attachment_id");
-	
-			
-?>
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-<link href="style.css" rel="stylesheet" type="text/css" />
-<link href="css/jquery.cluetip.css" rel="stylesheet" type="text/css" />
-<script src="jquery.min.js"></script>
-<script src="abcjs_editor_1.7-min.js" type="text/javascript"></script>
-<script src="js/jquery.cluetip.js"></script>
-<script src="js/general.js"></script>
-<script>
-	$(document).ready(function() {
-		$('a.jt').cluetip({
-		  cluetipClass: 'jtip',
-		  width: 350,
-		  arrows: true,
-		  dropShadow: false,
-		  hoverIntent: false,
-		  sticky: true,
-		  mouseOutClose: true,
-		  closePosition: 'title',
-		  closeText: '<img src="includes/images/cross.png" alt="close" />'
-		});
-	});
-	
-  function tgldiv(divname){
-	$(document).ready(function(){
-		$("#"+divname).slideToggle("slow");
-	  });
-  };
-  
-  <? 
-		if ( in_array('IS_ADMIN',$user_rights) ) {
-			?>
-     function removeUserTopic() {
-		$(document).ready(function(){
-			$('#t_user option:selected').each(function(){
-				$("<option/>").val($(this).val()).text($(this).text()).prependTo("#ex_users");
-				$("#t_user option[value="+$(this).val()+"]").remove();
-			});
-		});
-	}
-	function addUserTopic() {
-		$(document).ready(function(){
-			$('#ex_users option:selected').each(function(){
-				$("<option/>").val($(this).val()).text($(this).text()).prependTo("#t_user");
-				$("#ex_users option[value="+$(this).val()+"]").remove();
-			});
-		});
-	}
-	function removeGroupTopic() {
-		$(document).ready(function(){
-			$('#t_group option:selected').each(function(){
-				$("<option/>").val($(this).val()).text($(this).text()).prependTo("#ex_groups");
-				$("#t_group option[value="+$(this).val()+"]").remove();
-			});
-		});
-	}
-	function addGroupTopic() {
-		$(document).ready(function(){
-			$('#ex_groups option:selected').each(function(){
-				$("<option/>").val($(this).val()).text($(this).text()).prependTo("#t_group");
-				$("#ex_groups option[value="+$(this).val()+"]").remove();
-			});
-	  });
-	}
-	function selectAllOptions(selStr)
-	{
-	  var selObj = document.getElementById(selStr);
-	  for (var i=0; i<selObj.options.length; i++) {
-		selObj.options[i].selected = true;
-	  }
-	}
-	<? } ?>
-  </script>
-</head>
-<body>
-<div id="linksoben"><a href="topic.php?id=<?=$topicid?>#<?=$replyid?>"><?=TOPIC_TEXT_BACK?></a></div><? createRightHeader() ?><br>
-<br><br>
-
-<div id="panel-header"><?=TOPIC_EDIT_REPLY?></div>
-    <div id="editReply<?=$replyid?>" style="margin-left:20px; margin-right:20px;border:thin;border-style:dashed;padding: 10px;"><form action="<?=$_SERVER['PHP_SELF']?>" method="post" enctype="multipart/form-data"><input type="hidden" value="<?=$replyid?>" name="id"><div style="float:left"><textarea name="text" id="newReplyText<?=$replyid?>" rows="10" cols="50"><?=$res['reply_text']?></textarea></div><div style="float:left"><span class="formInfo"><a rel="faq.php?fid=1" href="faq.php?fid=1&width=475" class="jt" title="<?=TOPIC_TEXT_RULES?>">?</a></span></div><div style="clear:both"></div><?
-    
-	if ( in_array('CREATE_SHEETS', $user_rights) ) {
-		if ( $db->num_rows($query_m))
-			$db->data_seek($query_m,0);
-		while ( $r = $db->fetch_array($query_m)) {
-		?>
-        <div id="abc<?=$r['rm_id']?>Check" style="display:none"><pre id="abc<?=$r['rm_id']?>CheckText" class="abcCheckText"></pre></div>
-      <div style="float:left">  <textarea name="old_abc[<?=$r['music_id']?>]" id="abc<?=$r['rm_id']?>" cols="80" rows="15"><?=$r['music_text']?></textarea></div><div style="float:left"><span class="formInfo"><a rel="faq.php?fid=3" href="faq.php?fid=3" class="jt" title="<?=TOPIC_SHEET_INFO?>">?</a></span></div><div style="clear:both"></div>
-<input type="button" value="Check syntax" onclick="checkAbcSyntax('abc<?=$r['rm_id']?>')" /><div id="midi<?=$r['rm_id']?>"></div>
-<div id="warnings<?=$r['rm_id']?>"></div>
-<div id="music<?=$r['rm_id']?>"></div>
-<div id="paper0<?=$r['rm_id']?>"></div>
-
-        <?
-		}
-		showCreateSheet();
-	}
-	
-	if ( in_array('CREATE_ATTACHMENTS', $user_rights) ) {
-		echo TOPIC_TEXT_ATTACHMENTS.":";
-		if ( $db->num_rows($query_a) ) {
-			echo "<ul>";
-			while ( $r = $db->fetch_array($query_a) ) {
-				echo "<li>".$r['attachment_filename']." (<a href='delete_reply.php?aid=".$r['ra_id']."&id=".$replyid."'>".TOPIC_DELETE."</a>)</li>"; 	
+			$attachments = array();
+			while($r = $db->fetch_array($query_m)) {
+				$attachments[] = $r;
 			}
-			echo "</ul><br>";
-		}
-		echo '<br><input type="file" name="t_file"><span class="formInfo"><a href="faq.php?fid=2&width=475" rel="faq.php?fid=2&width=475" class="jt" title="'.TOPIC_TEXT_RULES_ATTACHMENT.'">?</a></span>';	
-	}
-	echo '<br><div style="float:left"><input type="submit" value="'.TOPIC_TEXT_SAVE.'"></div>';
+			$reply['attachments'] = $attachments;
 	
-	?><div style="float:left"><span class="formInfo"><a rel="faq.php?fid=4" href="faq.php?fid=4" class="jt" title="<?=TOPIC_TEXT_RULES?>">?</a></span></div><div style="clear:both"></div></form></div>
-    <script type="text/javascript">
-	window.onload = function() {
-		<?
-		if ( in_array('CREATE_SHEETS', $user_rights) ) {
-			if ( $db->num_rows($query_m))
-				$db->data_seek($query_m,0);
-			while ( $r = $db->fetch_array($query_m))
-				echo 'abc_editor'.$r['rm_id'].' = new ABCJS.Editor("abc'.$r['rm_id'].'", { paper_id: "paper0'.$r['rm_id'].'", midi_id:"midi'.$r['rm_id'].'", warnings_id:"warnings'.$r['rm_id'].'"});';
-		?>
+		} else {
+			$msgs[] = array('state' => 'nok', 'text' => DELETE_REPLY_NO_RIGHTS);
+		}
 		
-		abc_editor = new ABCJS.Editor("abc", { paper_id: "paper0", midi_id:"midi", warnings_id:"warnings" });
-	}
-	</script>
-    
-		<? 
 		if ( in_array('IS_ADMIN',$user_rights) ) {
-			?>
-       <br>
-     <div id="panel-header"><?=TOPIC_EDIT_TOPIC_SETTINGS?></div>
-	<div style="border:thin;border-style:dashed;padding:10px;margin-left:20px; margin-right:20px;">
-<form action="<?=$_SERVER['PHP_SELF']?>" method="post" enctype="multipart/form-data"  onsubmit="selectAllOptions('t_user');selectAllOptions('t_group');">
-Users: 
-<? 
-	$query_user = $db->query("select * from elo_user order by user_name");
-		
-	$users = array();
-	while ( $res = $db->fetch_array($query_user) )
-		$users[] = $res;
-		
-	$query_groups = $db->query("select * from elo_group order by group_name");
+				
+			$query_user = $db->query("select u.user_id, u.user_name, ut.tu_id from elo_user as u left join elo_topic_user as ut on (ut.user_id=u.user_id and ut.topic_id='".$topicid."') order by user_name");
+			
+			$users = array();
+			while ( $res = $db->fetch_array($query_user) )
+				$users[] = array('user_id' => $res['user_id'], 'user_name' => $res['user_name'], 'selected' => $res['tu_id'] ? 1 : 0);
+			$twig_data['users'] = $users;
 
-	$groups = array();
-	while ( $res = $db->fetch_array($query_groups) )
-		$groups[] = $res;
-	
-	$query_see_users = $db->query("select u.user_id from elo_user as u, elo_topic_user as ut where ut.user_id=u.user_id and ut.topic_id='".$topicid."'");
-	$query_see_groups = $db->query("select u.group_id from elo_group as u, elo_topic_group as ut where ut.group_id=u.group_id and ut.topic_id='".$topicid."'");
+			$query_groups = $db->query("select u.group_id, u.group_name, ut.tg_id from elo_group as u left join elo_topic_group as ut on (ut.group_id=u.group_id and ut.topic_id='".$topicid."' )order by group_name");
 
-	$user_topic = array();
-	while ( $r = $db->fetch_array($query_see_users) )
-		$user_topic[] = $r['user_id'];
-		
-	$group_topic = array();
-	while ( $r = $db->fetch_array($query_see_groups) )
-		$group_topic[] = $r['group_id'];
-	
-	$user_select = "";
-	$user_select_in = "";
-	foreach ( $users as $u ) {
-		if ( !in_array($u['user_id'],$user_topic) ) { 
-			$user_select .= "<option value='".$u['user_id']."'>".$u['user_name']."</option>";	
-		} else {
-			$user_select_in .= "<option value='".$u['user_id']."'>".$u['user_name']."</option>";	
+			$groups = array();
+			while ( $res = $db->fetch_array($query_groups) )
+				$groups[] = array('group_id' => $res['group_id'], 'group_name' => $res['group_name'], 'selected' => $res['tg_id']);
+
 		}
 	}
-	echo "<table><tr><td>Existing user:</td><td></td><td>User for this topic:</td></tr><tr><td><select id='ex_users' size='5' multiple='multiple'>".$user_select."</select></td>";
-	echo '<td><input type="button" value=">>" onClick="javascript:addUserTopic();"><br><input type="button" value="<<" onClick="javascript:removeUserTopic();"></td>';
-	echo "<td><select id='t_user' size='5' multiple='multiple' name='t_user[]'>".$user_select_in."</select></td></tr></table>";
-	
-
-?><br />
-Groups:
-<? 
-	$select_group = "";
-	$select_group_in = "";
-	foreach ( $groups as $g ) {
-		if ( !in_array( $g['group_id'],$group_topic ) ) {
-			$select_group .= "<option value='".$g['group_id']."'>".$g['group_name']."</option>";	
-		} else {
-			$select_group_in .= "<option value='".$g['group_id']."'>".$g['group_name']."</option>";	
-		}
-	}
-	
-	echo "<table><tr><td>Existing groups:</td><td></td><td>Groups for this topic:</td></tr><tr><td><select id='ex_groups' size='5' multiple='multiple'>".$select_group."</select></td>";
-	echo '<td><input type="button" value=">>" onClick="javascript:addGroupTopic();"><br><input type="button" value="<<" onClick="javascript:removeGroupTopic();"></td>';
-	echo "<td><select id='t_group' size='5' multiple='multiple' name='t_group[]'>".$select_group_in."</select></td></tr></table>";
-	
-?>
-<br />
-Title: <input type="text" name="t_topic_title" value="<?=$db->query_one("select topic_title from elo_topic where topic_id='".$topicid."'")?>"/><br>
-<input type="hidden" value="<?=$replyid?>" name="id">
-    <input type="submit" value="<?=TOPIC_TEXT_SAVE?>" name="edit_topic_admin" />
-</form>
-</div>
-	<?
-		}
-	}
-
-		} else {
-			$error = DELETE_REPLY_NO_RIGHTS;
-		}
-	} else {
-		$error = DELETE_REPLY_NO_RIGHTS;
-	}
-} else {
-	$error = DELETE_REPLY_NO_ID;	
 }
 
-?><br>
-<br>
-</body></html>
+$breadcrumb[] = array('href' => 'topic.php', 'text' => 'Topics');
+$breadcrumb[] = array('href' => 'topic.php?id='.$topicid."#".$replyid, 'text' => 'Topic');
+$twig_data['reply'] = $reply;
+$twig_data['msgs'] = $msgs;
+$twig_data['breadcrumb'] = $breadcrumb;
+echo $twig->render("edit-reply.twig", $twig_data);
